// types.d.ts
type mixed = unknown;
export type MayBe<T> = T | null | undefined; 
export type LangFunction = (s: string) => string;
// ...
export interface Ref {
    clearValidation(): void,
    readonly isValid: boolean,
    validate(): boolean,
    validateAsync(): Promise<boolean>,
}

export interface ContextRef extends Ref {
    getErrorMessage(name: string): string | void,
    getInput(name: string): InputRef | void,
    refreshMessage(): void,
    setErrorMessage(name: string, message: string): void,
}

export interface InputRef extends Ref {
    focus?: () => mixed,
    getErrorMessage: () => string,
    readonly name?: string,
    setErrorMessage(message: string): void,
}

export type RefObject<T> = {current: T | null, [prop: string]: mixed};
export type RefProp<T> = ((ref: T | null) => mixed) | RefObject<T>;

type TAsyncFailMessage = {
    Default: 0,
    CaughtError: 1,
};
export declare const AsyncFailMessage: TAsyncFailMessage;